refactor code to ensure datastax PreparedStatements are created only once
