+++ blueflood-core/src/integration-test/java/com/rackspacecloud/blueflood/io/IntegrationTestBase.java
@@ -20,6 +20,7 @@ import com.google.common.cache.Cache;
 import com.rackspacecloud.blueflood.cache.MetadataCache;
 import com.rackspacecloud.blueflood.io.astyanax.AstyanaxWriter;
 import com.rackspacecloud.blueflood.io.datastax.DCassandraUtilsIO;
+import com.rackspacecloud.blueflood.io.datastax.DLocatorIO;
 import com.rackspacecloud.blueflood.io.datastax.DPreaggregatedMetricsRW;
 import com.rackspacecloud.blueflood.outputs.formats.MetricData;
 import com.rackspacecloud.blueflood.rollup.Granularity;
@@ -121,17 +122,6 @@ public class IntegrationTestBase {
 
         Cache<String, Boolean> insertedLocators = (Cache<String, Boolean>) Whitebox.getInternalState(AstyanaxWriter.getInstance(), "insertedLocators");
         insertedLocators.invalidateAll();
-
-        return metric;
-    }
-
-    protected IMetric datastaxWriteEnumMetric(String name, String tenantid) throws Exception {
-        final List<IMetric> metrics = new ArrayList<IMetric>();
-        PreaggregatedMetric metric = getEnumMetric(name, tenantid, System.currentTimeMillis());
-        metrics.add(metric);
-
-        DPreaggregatedMetricsRW metricsRW = new DPreaggregatedMetricsRW();
-        metricsRW.insertMetrics(metrics, Granularity.FULL);
 
         return metric;
     }
