+++ blueflood-core/src/main/java/com/rackspacecloud/blueflood/io/datastax/DAbstractMetricIO.java
@@ -36,193 +36,15 @@ import java.util.*;
  */
 public abstract class DAbstractMetricIO {
 
-    /**
-     * The key of the metrics_preaggregated_* Column Families
-     */
-    public static final String KEY = "key";
-
-    /**
-     * The name of the first column
-     */
-    public static final String COLUMN1 = "column1";
-
-    /**
-     * The name of the value column
-     */
-    public static final String VALUE = "value";
-
-    protected static final String INSERT_KEY_COLUMN_VALUE_FORMAT = "INSERT INTO %s (key, column1, value) VALUES (?, ?, ?)";
-    protected static final String INSERT_KEY_COLUMN_VALUE_TTL_FORMAT = "INSERT INTO %s (key, column1, value) VALUES (?, ?, ?) USING TTL ?";
-    protected static final String SELECT_FOR_KEY = "SELECT * FROM %s WHERE key = :locator";
-    protected static final String SELECT_FOR_KEY_RANGE_FORMAT = "SELECT * FROM %s WHERE key = :locator AND column1 >= :tsStart AND column1 <= :tsEnd";
-
     private static final Logger LOG = LoggerFactory.getLogger(DAbstractMetricIO.class);
 
-    protected final PreparedStatement insertToMetricsPreaggrFullStatement;
-    protected final PreparedStatement insertToMetricsPreaggr5MStatement;
-    protected final PreparedStatement insertToMetricsPreaggr20MStatement;
-    protected final PreparedStatement insertToMetricsPreaggr60MStatement;
-    protected final PreparedStatement insertToMetricsPreaggr240MStatement;
-    protected final PreparedStatement insertToMetricsPreaggr1440MStatement;
-
-    private final PreparedStatement selectFromMetricsPreaggrFullForRangeStatement;
-    private final PreparedStatement selectFromMetricsPreaggr5MForRangeStatement;
-    private final PreparedStatement selectFromMetricsPreaggr20MForRangeStatement;
-    private final PreparedStatement selectFromMetricsPreaggr60MForRangeStatement;
-    private final PreparedStatement selectFromMetricsPreaggr240MForRangeStatement;
-    private final PreparedStatement selectFromMetricsPreaggr1440MForRangeStatement;
-
-    protected final PreparedStatement insertToMetricsBasicFullStatement;
-    protected final PreparedStatement insertToMetricsBasic5MStatement;
-    protected final PreparedStatement insertToMetricsBasic20MStatement;
-    protected final PreparedStatement insertToMetricsBasic60MStatement;
-    protected final PreparedStatement insertToMetricsBasic240MStatement;
-    protected final PreparedStatement insertToMetricsBasic1440MStatement;
-
-    private final PreparedStatement selectFromMetricsStringForRangeStatement;
-
-    private final PreparedStatement selectFromMetricsBasicFullForRangeStatement;
-    private final PreparedStatement selectFromMetricsBasic5MForRangeStatement;
-    private final PreparedStatement selectFromMetricsBasic20MForRangeStatement;
-    private final PreparedStatement selectFromMetricsBasic60MForRangeStatement;
-    private final PreparedStatement selectFromMetricsBasic240MForRangeStatement;
-    private final PreparedStatement selectFromMetricsBasic1440MForRangeStatement;
-
     protected Session session;
-    protected Map<String, PreparedStatement> cfNameToSelectStatement;
-    protected Map<Granularity, PreparedStatement> preaggrGranToInsertStatement;
-    protected Map<Granularity, PreparedStatement> basicGranToInsertStatement;
+
+    protected final DMetricsCFPreparedStatements metricsCFPreparedStatements;
 
     protected DAbstractMetricIO() {
+        metricsCFPreparedStatements = DMetricsCFPreparedStatements.getInstance();
         session = DatastaxIO.getSession();
-
-        //
-        // Preaggr insert statements
-        //
-        insertToMetricsPreaggrFullStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_FULL_NAME));
-        insertToMetricsPreaggr5MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_5M_NAME));
-        insertToMetricsPreaggr20MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_20M_NAME));
-        insertToMetricsPreaggr60MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_60M_NAME));
-        insertToMetricsPreaggr240MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_240M_NAME));
-        insertToMetricsPreaggr1440MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_1440M_NAME));
-
-        //
-        // Preaggr select statements
-        //
-        selectFromMetricsPreaggrFullForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_FULL_NAME));
-        selectFromMetricsPreaggr5MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_5M_NAME));
-        selectFromMetricsPreaggr20MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_20M_NAME));
-        selectFromMetricsPreaggr60MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_60M_NAME));
-        selectFromMetricsPreaggr240MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_240M_NAME));
-        selectFromMetricsPreaggr1440MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_PREAGGREGATED_1440M_NAME));
-
-        cfNameToSelectStatement = new HashMap<String, PreparedStatement>() {{
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_FULL_NAME, selectFromMetricsPreaggrFullForRangeStatement);
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_5M_NAME, selectFromMetricsPreaggr5MForRangeStatement);
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_20M_NAME, selectFromMetricsPreaggr20MForRangeStatement);
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_60M_NAME, selectFromMetricsPreaggr60MForRangeStatement);
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_240M_NAME, selectFromMetricsPreaggr240MForRangeStatement);
-            put(CassandraModel.CF_METRICS_PREAGGREGATED_1440M_NAME, selectFromMetricsPreaggr1440MForRangeStatement);
-        }};
-
-        preaggrGranToInsertStatement = new HashMap<Granularity, PreparedStatement>() {{
-            put(Granularity.FULL, insertToMetricsPreaggrFullStatement);
-            put(Granularity.MIN_5, insertToMetricsPreaggr5MStatement);
-            put(Granularity.MIN_20, insertToMetricsPreaggr20MStatement);
-            put(Granularity.MIN_60, insertToMetricsPreaggr60MStatement);
-            put(Granularity.MIN_240, insertToMetricsPreaggr240MStatement);
-            put(Granularity.MIN_1440, insertToMetricsPreaggr1440MStatement);
-        }};
-
-        //
-        // Basic insert statements
-        //
-        insertToMetricsBasicFullStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_FULL_NAME));
-        insertToMetricsBasic5MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_5M_NAME));
-        insertToMetricsBasic20MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_20M_NAME));
-        insertToMetricsBasic60MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_60M_NAME));
-        insertToMetricsBasic240MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_240M_NAME));
-        insertToMetricsBasic1440MStatement = session.prepare(
-                String.format(INSERT_KEY_COLUMN_VALUE_TTL_FORMAT,
-                        CassandraModel.CF_METRICS_1440M_NAME));
-
-        //
-        // Basic select statements
-        //
-        selectFromMetricsStringForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_STRING_NAME));
-        selectFromMetricsBasicFullForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_FULL_NAME));
-        selectFromMetricsBasic5MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_5M_NAME));
-        selectFromMetricsBasic20MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_20M_NAME));
-        selectFromMetricsBasic60MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_60M_NAME));
-        selectFromMetricsBasic240MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_240M_NAME));
-        selectFromMetricsBasic1440MForRangeStatement = session.prepare(
-                String.format(SELECT_FOR_KEY_RANGE_FORMAT,
-                        CassandraModel.CF_METRICS_1440M_NAME));
-
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_STRING_NAME, selectFromMetricsStringForRangeStatement );
-
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_FULL_NAME, selectFromMetricsBasicFullForRangeStatement );
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_5M_NAME, selectFromMetricsBasic5MForRangeStatement );
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_20M_NAME, selectFromMetricsBasic20MForRangeStatement );
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_60M_NAME, selectFromMetricsBasic60MForRangeStatement );
-        cfNameToSelectStatement.put( CassandraModel.CF_METRICS_240M_NAME, selectFromMetricsBasic240MForRangeStatement );
-        cfNameToSelectStatement.put(CassandraModel.CF_METRICS_1440M_NAME, selectFromMetricsBasic1440MForRangeStatement);
-
-        basicGranToInsertStatement = new HashMap<Granularity, PreparedStatement>() {{
-            // NOTE:  this shoudn't be called.  explain why later
-            put(Granularity.FULL, insertToMetricsBasicFullStatement );
-            put(Granularity.MIN_5, insertToMetricsBasic5MStatement);
-            put(Granularity.MIN_20, insertToMetricsBasic20MStatement);
-            put(Granularity.MIN_60, insertToMetricsBasic60MStatement);
-            put(Granularity.MIN_240, insertToMetricsBasic240MStatement);
-            put(Granularity.MIN_1440, insertToMetricsBasic1440MStatement);
-        }};
     }
 
     /**
@@ -335,10 +157,10 @@ public abstract class DAbstractMetricIO {
             // Strings and Booleans don't get rolled up.  I'd like to verify
             // that none are passed in, but that would require a db access
 
-            statement = basicGranToInsertStatement.get( granularity );
+            statement = metricsCFPreparedStatements.basicGranToInsertStatement.get( granularity );
         }
         else {
-            statement = preaggrGranToInsertStatement.get(granularity);
+            statement = metricsCFPreparedStatements.preaggrGranToInsertStatement.get(granularity);
         }
 
         BoundStatement bound = statement.bind(locator.toString(),
@@ -386,7 +208,7 @@ public abstract class DAbstractMetricIO {
      */
     protected List<ResultSetFuture> selectForLocatorAndRange(String columnFamily, Locator locator, Range range) {
         List<ResultSetFuture> resultsFutures = new ArrayList<ResultSetFuture>();
-        PreparedStatement statement = cfNameToSelectStatement.get(columnFamily);
+        PreparedStatement statement = metricsCFPreparedStatements.cfNameToSelectStatement.get(columnFamily);
         resultsFutures.add(
                 session.executeAsync(statement.bind(locator.toString(), range.getStart(), range.getStop())));
         return resultsFutures;
@@ -405,10 +227,10 @@ public abstract class DAbstractMetricIO {
             try {
                 List<Row> rows = future.getUninterruptibly().all();
                 for (Row row : rows) {
-                    String key = row.getString(KEY);
+                    String key = row.getString(DMetricsCFPreparedStatements.KEY);
                     Locator loc = Locator.createLocatorFromDbKey(key);
-                    Long hash = row.getLong(COLUMN1);
-                    locatorTimestampRollup.put(loc, hash, (T)fromByteBuffer(row.getBytes(VALUE)));
+                    Long hash = row.getLong(DMetricsCFPreparedStatements.COLUMN1);
+                    locatorTimestampRollup.put(loc, hash, (T)fromByteBuffer(row.getBytes(DMetricsCFPreparedStatements.VALUE)));
                 }
             } catch (Exception ex) {
                 Instrumentation.markReadError();
